+++ front-end/src/channel/List.js
@@ -1,115 +1,154 @@
-import {forwardRef, useImperativeHandle, useLayoutEffect, useRef} from 'react'
+import {
+  forwardRef,
+  useImperativeHandle,
+  useLayoutEffect,
+  useRef,
+} from "react";
 /** @jsx jsx */
-import { jsx } from '@emotion/core'
+import { jsx } from "@emotion/core";
 // Layout
-import { useTheme } from '@material-ui/core/styles';
+import { useTheme, withStyles } from "@material-ui/core/styles";
 // Markdown
-import unified from 'unified'
-import markdown from 'remark-parse'
-import remark2rehype from 'remark-rehype'
-import html from 'rehype-stringify'
+import unified from "unified";
+import markdown from "remark-parse";
+import remark2rehype from "remark-rehype";
+import html from "rehype-stringify";
 // Time
-import dayjs from 'dayjs'
-import calendar from 'dayjs/plugin/calendar'
-import updateLocale from 'dayjs/plugin/updateLocale'
-dayjs.extend(calendar)
-dayjs.extend(updateLocale)
-dayjs.updateLocale('en', {
-  calendar: {
-    sameElse: 'DD/MM/YYYY hh:mm A'
-  }
-})
+import dayjs from "dayjs";
+import calendar from "dayjs/plugin/calendar";
+import updateLocale from "dayjs/plugin/updateLocale";
+import Button from "@material-ui/core/Button";
+import Tooltip from "@material-ui/core/Tooltip";
+import Grid from "@material-ui/core/Grid";
+import { orange } from "@material-ui/core/colors";
 
 const useStyles = (theme) => ({
   root: {
-    position: 'relative',
-    flex: '1 1 auto',
-    'pre': {
-      
-      overflowY: 'auto',
+    position: "relative",
+    flex: "1 1 auto",
+    border: "1px solid rgba(255, 255, 255, 0.2)",
+    pre: {
+      overflowY: "auto",
     },
-    '& ul': {
-      'margin': 0,
-      'padding': 0,
-      'textIndent': 0,
-      'listStyleType': 0,
+    "& ul": {
+      margin: 0,
+      padding: 0,
+      textIndent: 0,
+      listStyleType: 0,
     },
   },
+  title: {
+    margin: "0px",
+    padding: "0px 20px",
+    border: "solid 1px rgb(255, 255, 255, 0.2)",
+  },
   message: {
-    padding: '.2rem .5rem',
-    ':hover': {
-      backgroundColor: 'rgba(255,255,255,.05)',
+    padding: ".2rem .5rem",
+    ":hover": {
+      backgroundColor: "rgba(255,255,255,.05)",
     },
   },
   fabWrapper: {
-    position: 'absolute',
+    position: "absolute",
     right: 0,
     top: 0,
-    width: '50px',
+    width: "50px",
   },
   fab: {
-    position: 'fixed !important',
+    position: "fixed !important",
     top: 0,
-    width: '50px',
+    width: "50px",
   },
-})
+});
 
-export default forwardRef(({
-  channel,
-  messages,
-  onScrollDown,
-}, ref) => {
-  const styles = useStyles(useTheme())
+const ColorButton = withStyles((theme) => ({
+  root: {
+    color: "white",
+    backgroundColor: orange[500],
+    "&:hover": {
+      backgroundColor: orange[700],
+    },
+  },
+}))(Button);
+
+dayjs.extend(calendar);
+dayjs.extend(updateLocale);
+dayjs.updateLocale("en", {
+  calendar: {
+    sameElse: "DD/MM/YYYY hh:mm A",
+  },
+});
+
+export default forwardRef(({ channel, messages, onScrollDown }, ref) => {
+  const styles = useStyles(useTheme());
   // Expose the `scroll` action
   useImperativeHandle(ref, () => ({
-    scroll: scroll
+    scroll: scroll,
   }));
-  const rootEl = useRef(null)
-  const scrollEl = useRef(null)
+  const rootEl = useRef(null);
+  const scrollEl = useRef(null);
   const scroll = () => {
-    scrollEl.current.scrollIntoView()
-  }
+    scrollEl.current.scrollIntoView();
+  };
   // See https://dev.to/n8tb1t/tracking-scroll-position-with-react-hooks-3bbj
-  const throttleTimeout = useRef(null) // react-hooks/exhaustive-deps
-  useLayoutEffect( () => {
-    const rootNode = rootEl.current // react-hooks/exhaustive-deps
+  const throttleTimeout = useRef(null); // react-hooks/exhaustive-deps
+  useLayoutEffect(() => {
+    const rootNode = rootEl.current; // react-hooks/exhaustive-deps
     const handleScroll = () => {
       if (throttleTimeout.current === null) {
         throttleTimeout.current = setTimeout(() => {
-          throttleTimeout.current = null
-          const {scrollTop, offsetHeight, scrollHeight} = rootNode // react-hooks/exhaustive-deps
-          onScrollDown(scrollTop + offsetHeight < scrollHeight)
-        }, 200)
+          throttleTimeout.current = null;
+          const { scrollTop, offsetHeight, scrollHeight } = rootNode; // react-hooks/exhaustive-deps
+          onScrollDown(scrollTop + offsetHeight < scrollHeight);
+        }, 200);
       }
-    }
-    handleScroll()
-    rootNode.addEventListener('scroll', handleScroll)
-    return () => rootNode.removeEventListener('scroll', handleScroll)
-  })
+    };
+    handleScroll();
+    rootNode.addEventListener("scroll", handleScroll);
+    return () => rootNode.removeEventListener("scroll", handleScroll);
+  });
+
+  
+
   return (
     <div css={styles.root} ref={rootEl}>
-      <h1>Messages for {channel.name}</h1>
+      <Grid
+        container
+        direction="row"
+        justify="space-between"
+        alignItems="center"
+        css={styles.title}
+      >
+        <Grid item>
+          <h1>{channel.name}</h1>
+        </Grid>
+        <Grid item>
+          <Tooltip title="members" arrow>
+            <ColorButton>Channel Members</ColorButton>
+          </Tooltip>
+        </Grid>
+      </Grid>
+
       <ul>
-        { messages.map( (message, i) => {
-            const {contents: content} = unified()
+        {messages.map((message, i) => {
+          const { contents: content } = unified()
             .use(markdown)
             .use(remark2rehype)
             .use(html)
-            .processSync(message.content)
-            return (
-              <li key={i} css={styles.message}>
-                <p>
-                  <span>{message.author}</span>
-                  {' - '}
-                  <span>{dayjs().calendar(message.creation)}</span>
-                </p>
-                <div dangerouslySetInnerHTML={{__html: content}}>
-                </div>
-              </li>
-            )
+            .processSync(message.content);
+          return (
+            <li key={i} css={styles.message}>
+              <p>
+                <span>{message.author}</span>
+                {" - "}
+                <span>{dayjs().calendar(message.creation)}</span>
+              </p>
+              <div dangerouslySetInnerHTML={{ __html: content }}></div>
+            </li>
+          );
         })}
       </ul>
       <div ref={scrollEl} />
     </div>
-  )
-})
+  );
+});